Updating All Project
We updated our project and we changed the some parts of our codes. Before that we could not reach the additional features. Now at least we can reach our additional features. We decided that we should change our method for taking additional features. In first we decided that additional feature is a weak entity but now it is an entity. Moreover before that it has relationship with product entity but now it is directly going to product entity with many to many relationship.